+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **react** - Allow globally configuring the React Query key via `configureQueryKey`. Export `DEFAULT_QUERY_KEY`.
 - **core** - Add a query builder.
+- **react** - `useLdhopQuery` also accepts `LdhopQueryBuilder`.
 
 ## [1.0.0-alpha.5] - 2025-12-30